getIconInfo, getImageInfo の実装を移動
